formatted codes with forge fmt

+++ src/JudgeToken.sol
@@ -9,21 +9,12 @@ import {ERC20Capped} from "../lib/openzeppelin-contracts/contracts/token/ERC20/e
 import {ERC20Votes} from "../lib/openzeppelin-contracts/contracts/token/ERC20/extensions/ERC20Votes.sol";
 import {Nonces} from "../lib/openzeppelin-contracts/contracts/utils/Nonces.sol";
 
-contract JudgeToken is
-    ERC20,
-    ERC20Burnable,
-    ERC20Permit,
-    ERC20Votes,
-    AccessControl,
-    ERC20Capped
-{
+contract JudgeToken is ERC20, ERC20Burnable, ERC20Permit, ERC20Votes, AccessControl, ERC20Capped {
     bytes32 public constant MINTER_ROLE = keccak256("MINTER_ROLE");
 
-    event Minted(address indexed caller, address indexed to, uint amount);
+    event Minted(address indexed caller, address indexed to, uint256 amount);
 
-    constructor(
-        uint256 initialSupply
-    )
+    constructor(uint256 initialSupply)
         ERC20("JudgeToken", "JUDGE")
         ERC20Capped(100_000_000 * 10 ** decimals())
         ERC20Permit("JudgeToken")
@@ -34,15 +25,12 @@ contract JudgeToken is
         emit Minted(msg.sender, msg.sender, initialSupply);
     }
 
-    function mint(address to, uint amount) external onlyRole(MINTER_ROLE) {
+    function mint(address to, uint256 amount) external onlyRole(MINTER_ROLE) {
         _mint(to, amount);
         emit Minted(msg.sender, to, amount);
     }
 
-    function setRoleAdmin(
-        bytes32 role,
-        bytes32 adminRole
-    ) external onlyRole(DEFAULT_ADMIN_ROLE) {
+    function setRoleAdmin(bytes32 role, bytes32 adminRole) external onlyRole(DEFAULT_ADMIN_ROLE) {
         _setRoleAdmin(role, adminRole);
     }
 
@@ -50,17 +38,11 @@ contract JudgeToken is
         return _getTotalSupply();
     }
 
-    function _update(
-        address from,
-        address to,
-        uint value
-    ) internal override(ERC20, ERC20Capped, ERC20Votes) {
+    function _update(address from, address to, uint256 value) internal override(ERC20, ERC20Capped, ERC20Votes) {
         super._update(from, to, value);
     }
 
-    function nonces(
-        address owner
-    ) public view override(ERC20Permit, Nonces) returns (uint256) {
+    function nonces(address owner) public view override(ERC20Permit, Nonces) returns (uint256) {
         return super.nonces(owner);
     }
 }
