+++ test/JudgeToken.t.sol
@@ -25,16 +25,8 @@ contract JudgeTokenTest is Test {
     error ERC20InvalidReceiver(address receiver);
     error ERC20InvalidSpender(address spender);
     error ERC20ExceededCap(uint256 increasedSupply, uint256 cap);
-    error ERC20InsufficientAllowance(
-        address spender,
-        uint256 allowance,
-        uint256 needed
-    );
-    error ERC20InsufficientBalance(
-        address sender,
-        uint256 balance,
-        uint256 needed
-    );
+    error ERC20InsufficientAllowance(address spender, uint256 allowance, uint256 needed);
+    error ERC20InsufficientBalance(address sender, uint256 balance, uint256 needed);
     error AccessControlBadConfirmation();
     error ERC2612ExpiredSignature(uint256 deadline);
     error ERC2612InvalidSigner(address signer, address owner);
@@ -65,13 +57,7 @@ contract JudgeTokenTest is Test {
         uint256 mintAmount = 99_000_001 * 10 ** uint256(decimals);
         uint256 increasedSupply = mintAmount + initialSupply;
         assertEq(cap, judgeToken.cap());
-        vm.expectRevert(
-            abi.encodeWithSelector(
-                ERC20ExceededCap.selector,
-                increasedSupply,
-                judgeToken.cap()
-            )
-        );
+        vm.expectRevert(abi.encodeWithSelector(ERC20ExceededCap.selector, increasedSupply, judgeToken.cap()));
         judgeToken.mint(user1, mintAmount);
     }
 
@@ -94,24 +80,13 @@ contract JudgeTokenTest is Test {
         bytes32 minterRole = judgeToken.MINTER_ROLE();
         judgeToken.mint(user1, mintAmount);
         assertEq(judgeToken.balanceOf(user1), mintAmount);
-        assertEq(
-            judgeToken.totalSupply(),
-            judgeToken.balanceOf(user1) + judgeToken.balanceOf(owner)
-        );
+        assertEq(judgeToken.totalSupply(), judgeToken.balanceOf(user1) + judgeToken.balanceOf(owner));
 
         vm.prank(user1);
-        vm.expectRevert(
-            abi.encodeWithSelector(
-                AccessControlUnauthorizedAccount.selector,
-                user1,
-                minterRole
-            )
-        );
+        vm.expectRevert(abi.encodeWithSelector(AccessControlUnauthorizedAccount.selector, user1, minterRole));
         judgeToken.mint(user1, mintAmount);
 
-        vm.expectRevert(
-            abi.encodeWithSelector(ERC20InvalidReceiver.selector, address(0))
-        );
+        vm.expectRevert(abi.encodeWithSelector(ERC20InvalidReceiver.selector, address(0)));
         judgeToken.mint(zeroAddress, mintAmount);
     }
 
@@ -123,9 +98,7 @@ contract JudgeTokenTest is Test {
         judgeToken.burn(burnAmount);
         assertEq(mintAmount - burnAmount, judgeToken.balanceOf(user1));
 
-        vm.expectRevert(
-            abi.encodeWithSelector(ERC20InvalidSender.selector, address(0))
-        );
+        vm.expectRevert(abi.encodeWithSelector(ERC20InvalidSender.selector, address(0)));
         vm.prank(zeroAddress);
         judgeToken.burn(burnAmount);
     }
@@ -146,14 +119,7 @@ contract JudgeTokenTest is Test {
         assertEq(judgeToken.balanceOf(user2), balanceLeft);
 
         // This will revert with insufficiant allowance before reverting with invalid sender
-        vm.expectRevert(
-            abi.encodeWithSelector(
-                ERC20InsufficientAllowance.selector,
-                user1,
-                0,
-                burnAmount
-            )
-        );
+        vm.expectRevert(abi.encodeWithSelector(ERC20InsufficientAllowance.selector, user1, 0, burnAmount));
         vm.prank(user1);
         judgeToken.burnFrom(zeroAddress, burnAmount);
     }
@@ -177,24 +143,17 @@ contract JudgeTokenTest is Test {
 
         vm.expectRevert(
             abi.encodeWithSelector(
-                ERC20InsufficientBalance.selector,
-                user2,
-                judgeToken.balanceOf(user2),
-                amountGreaterThanBalance
+                ERC20InsufficientBalance.selector, user2, judgeToken.balanceOf(user2), amountGreaterThanBalance
             )
         );
         vm.prank(user2);
         judgeToken.transfer(user1, amountGreaterThanBalance);
 
-        vm.expectRevert(
-            abi.encodeWithSelector(ERC20InvalidSender.selector, address(0))
-        );
+        vm.expectRevert(abi.encodeWithSelector(ERC20InvalidSender.selector, address(0)));
         vm.prank(zeroAddress);
         judgeToken.transfer(user1, amount);
 
-        vm.expectRevert(
-            abi.encodeWithSelector(ERC20InvalidReceiver.selector, address(0))
-        );
+        vm.expectRevert(abi.encodeWithSelector(ERC20InvalidReceiver.selector, address(0)));
         vm.prank(user2);
         judgeToken.transfer(zeroAddress, amount);
     }
@@ -209,12 +168,7 @@ contract JudgeTokenTest is Test {
         judgeToken.approve(user1, allowance);
 
         vm.expectRevert(
-            abi.encodeWithSelector(
-                ERC20InsufficientAllowance.selector,
-                user1,
-                allowance,
-                valueGreaterThanAllowance
-            )
+            abi.encodeWithSelector(ERC20InsufficientAllowance.selector, user1, allowance, valueGreaterThanAllowance)
         );
         vm.prank(user1);
         judgeToken.transferFrom(user2, user3, valueGreaterThanAllowance);
@@ -240,9 +194,7 @@ contract JudgeTokenTest is Test {
         uint256 allowance = 30_000 * 10 ** uint256(decimals);
         judgeToken.mint(user2, mintAmount);
 
-        vm.expectRevert(
-            abi.encodeWithSelector(ERC20InvalidSpender.selector, address(0))
-        );
+        vm.expectRevert(abi.encodeWithSelector(ERC20InvalidSpender.selector, address(0)));
         vm.prank(user2);
         judgeToken.approve(zeroAddress, allowance);
     }
@@ -254,13 +206,7 @@ contract JudgeTokenTest is Test {
         judgeToken.grantRole(minterRole, user1);
         assertTrue(judgeToken.hasRole(minterRole, user1));
 
-        vm.expectRevert(
-            abi.encodeWithSelector(
-                AccessControlUnauthorizedAccount.selector,
-                user1,
-                defaultAdmin
-            )
-        );
+        vm.expectRevert(abi.encodeWithSelector(AccessControlUnauthorizedAccount.selector, user1, defaultAdmin));
         vm.prank(user1);
         judgeToken.grantRole(minterRole, user2);
     }
@@ -284,13 +230,7 @@ contract JudgeTokenTest is Test {
         judgeToken.grantRole(minterRole, user2);
         assertTrue(judgeToken.hasRole(minterRole, user2));
 
-        vm.expectRevert(
-            abi.encodeWithSelector(
-                AccessControlUnauthorizedAccount.selector,
-                user1,
-                defaultAdmin
-            )
-        );
+        vm.expectRevert(abi.encodeWithSelector(AccessControlUnauthorizedAccount.selector, user1, defaultAdmin));
         vm.prank(user1);
         judgeToken.revokeRole(minterRole, user2);
         assertTrue(judgeToken.hasRole(minterRole, user2));
@@ -308,18 +248,14 @@ contract JudgeTokenTest is Test {
         judgeToken.renounceRole(minterRole, user2);
         assertFalse(judgeToken.hasRole(minterRole, user2));
 
-        vm.expectRevert(
-            abi.encodeWithSelector(AccessControlBadConfirmation.selector)
-        );
+        vm.expectRevert(abi.encodeWithSelector(AccessControlBadConfirmation.selector));
         vm.prank(user2);
         judgeToken.renounceRole(minterRole, user1);
     }
 
     function testSupportsInterface() public {
         assertTrue(judgeToken.supportsInterface(type(IERC165).interfaceId));
-        assertTrue(
-            judgeToken.supportsInterface(type(IAccessControl).interfaceId)
-        );
+        assertTrue(judgeToken.supportsInterface(type(IAccessControl).interfaceId));
     }
 
     function testDelegate() public {
@@ -347,7 +283,7 @@ contract JudgeTokenTest is Test {
         assertEq(judgeToken.getVotes(owner), initialSupply);
         uint256 timePoint = block.number;
 
-        for (uint i; i < 10; i++) {
+        for (uint256 i; i < 10; i++) {
             vm.roll(block.number + 1);
         }
         judgeToken.transfer(user1, amount);
@@ -359,7 +295,7 @@ contract JudgeTokenTest is Test {
     function testGetPastTotalSupply() public {
         uint256 timePoint1 = block.number;
         uint256 amount = 100_500 * 10 ** uint256(decimals);
-        for (uint i; i < 10; i++) {
+        for (uint256 i; i < 10; i++) {
             vm.roll(block.number + 1);
         }
         judgeToken.mint(user2, amount);
@@ -371,12 +307,12 @@ contract JudgeTokenTest is Test {
         uint256 amount = 100_000 * 10 ** uint256(decimals);
         judgeToken.delegate(owner);
 
-        for (uint i; i < 5; i++) {
+        for (uint256 i; i < 5; i++) {
             vm.roll(block.number + 1);
         }
         judgeToken.transfer(user1, amount);
 
-        for (uint i; i < 10; i++) {
+        for (uint256 i; i < 10; i++) {
             vm.roll(block.number + 1);
         }
         judgeToken.transfer(user2, amount);
@@ -388,17 +324,17 @@ contract JudgeTokenTest is Test {
         uint256 amount = 10_000 * 10 ** uint256(decimals);
         uint256 amount2 = 100_500 * 10 ** uint256(decimals);
         assertEq(judgeToken.numCheckpoints(owner), 0);
-        for (uint i; i < 10; i++) {
+        for (uint256 i; i < 10; i++) {
             vm.roll(block.number + 1);
         }
         judgeToken.delegate(owner);
 
-        for (uint i; i < 10; i++) {
+        for (uint256 i; i < 10; i++) {
             vm.roll(block.number + 1);
         }
         judgeToken.transfer(user1, amount);
 
-        for (uint i; i < 10; i++) {
+        for (uint256 i; i < 10; i++) {
             vm.roll(block.number + 1);
         }
         judgeToken.mint(owner, amount2);
@@ -406,29 +342,17 @@ contract JudgeTokenTest is Test {
         uint32 num = judgeToken.numCheckpoints(owner);
 
         for (uint32 i = 0; i < num; i++) {
-            Checkpoints.Checkpoint208 memory ckpt = judgeToken.checkpoints(
-                owner,
-                i
-            );
+            Checkpoints.Checkpoint208 memory ckpt = judgeToken.checkpoints(owner, i);
             console.log("Checkpoint", i);
             console.log("fromBlock", ckpt._key);
             console.log("Votes", ckpt._value);
         }
 
-        Checkpoints.Checkpoint208 memory ckpt1 = judgeToken.checkpoints(
-            owner,
-            0
-        );
+        Checkpoints.Checkpoint208 memory ckpt1 = judgeToken.checkpoints(owner, 0);
         assertEq(ckpt1._value, initialSupply);
-        Checkpoints.Checkpoint208 memory ckpt2 = judgeToken.checkpoints(
-            owner,
-            1
-        );
+        Checkpoints.Checkpoint208 memory ckpt2 = judgeToken.checkpoints(owner, 1);
         assertEq(ckpt2._value, initialSupply - amount);
-        Checkpoints.Checkpoint208 memory ckpt3 = judgeToken.checkpoints(
-            owner,
-            2
-        );
+        Checkpoints.Checkpoint208 memory ckpt3 = judgeToken.checkpoints(owner, 2);
         assertEq(ckpt3._value, initialSupply - amount + amount2);
     }
 
@@ -445,9 +369,7 @@ contract JudgeTokenTest is Test {
     ) internal pure returns (bytes32) {
         bytes32 DOMAIN_SEPARATOR = keccak256(
             abi.encode(
-                keccak256(
-                    "EIP712Domain(string name,string version,uint256 chainId,address verifyingContract)"
-                ),
+                keccak256("EIP712Domain(string name,string version,uint256 chainId,address verifyingContract)"),
                 keccak256(bytes(name)),
                 keccak256(bytes(version)),
                 chainId,
@@ -455,27 +377,16 @@ contract JudgeTokenTest is Test {
             )
         );
 
-        bytes32 PERMIT_TYPEHASH = keccak256(
-            "Permit(address owner,address spender,uint256 value,uint256 nonce,uint256 deadline)"
-        );
+        bytes32 PERMIT_TYPEHASH =
+            keccak256("Permit(address owner,address spender,uint256 value,uint256 nonce,uint256 deadline)");
 
-        return
-            keccak256(
-                abi.encodePacked(
-                    "\x19\x01",
-                    DOMAIN_SEPARATOR,
-                    keccak256(
-                        abi.encode(
-                            PERMIT_TYPEHASH,
-                            signer,
-                            spender,
-                            value,
-                            nonce,
-                            deadline
-                        )
-                    )
-                )
-            );
+        return keccak256(
+            abi.encodePacked(
+                "\x19\x01",
+                DOMAIN_SEPARATOR,
+                keccak256(abi.encode(PERMIT_TYPEHASH, signer, spender, value, nonce, deadline))
+            )
+        );
     }
 
     function testNonces() public {
@@ -484,15 +395,7 @@ contract JudgeTokenTest is Test {
         uint256 user1Nonce = judgeToken.nonces(user1);
         uint256 deadline = block.timestamp + 1 days;
         bytes32 digest = getPermitDigest(
-            "JudgeToken",
-            "1",
-            address(judgeToken),
-            block.chainid,
-            user1,
-            user2,
-            value,
-            user1Nonce,
-            deadline
+            "JudgeToken", "1", address(judgeToken), block.chainid, user1, user2, value, user1Nonce, deadline
         );
 
         (uint8 v, bytes32 r, bytes32 s) = vm.sign(pKey, digest);
@@ -506,17 +409,8 @@ contract JudgeTokenTest is Test {
         judgeToken.mint(user1, mintValue);
         uint256 nonce = judgeToken.nonces(user1);
         uint256 deadline = block.timestamp + 1 days;
-        bytes32 digest = getPermitDigest(
-            "JudgeToken",
-            "1",
-            address(judgeToken),
-            block.chainid,
-            user1,
-            user2,
-            value,
-            nonce,
-            deadline
-        );
+        bytes32 digest =
+            getPermitDigest("JudgeToken", "1", address(judgeToken), block.chainid, user1, user2, value, nonce, deadline);
 
         (uint8 v, bytes32 r, bytes32 s) = vm.sign(pKey, digest);
         judgeToken.permit(user1, user2, value, deadline, v, r, s);
@@ -535,23 +429,12 @@ contract JudgeTokenTest is Test {
         judgeToken.mint(user1, mintValue);
         uint256 nonce = judgeToken.nonces(user1);
         uint256 deadline = block.timestamp + 1 days;
-        bytes32 digest = getPermitDigest(
-            "JudgeToken",
-            "1",
-            address(judgeToken),
-            block.chainid,
-            user1,
-            user2,
-            value,
-            nonce,
-            deadline
-        );
+        bytes32 digest =
+            getPermitDigest("JudgeToken", "1", address(judgeToken), block.chainid, user1, user2, value, nonce, deadline);
 
         (uint8 v, bytes32 r, bytes32 s) = vm.sign(pKey, digest);
         vm.warp(deadline + 1 days);
-        vm.expectRevert(
-            abi.encodeWithSelector(ERC2612ExpiredSignature.selector, deadline)
-        );
+        vm.expectRevert(abi.encodeWithSelector(ERC2612ExpiredSignature.selector, deadline));
         judgeToken.permit(user1, user2, value, deadline, v, r, s);
     }
 
@@ -561,22 +444,11 @@ contract JudgeTokenTest is Test {
         judgeToken.mint(user1, mintValue);
         uint256 nonce = judgeToken.nonces(user1);
         uint256 deadline = block.timestamp + 1 days;
-        bytes32 digest = getPermitDigest(
-            "JudgeToken",
-            "1",
-            address(judgeToken),
-            block.chainid,
-            user1,
-            user2,
-            value,
-            nonce,
-            deadline
-        );
+        bytes32 digest =
+            getPermitDigest("JudgeToken", "1", address(judgeToken), block.chainid, user1, user2, value, nonce, deadline);
 
         (uint8 v, bytes32 r, bytes32 s) = vm.sign(pKey2, digest);
-        vm.expectRevert(
-            abi.encodeWithSelector(ERC2612InvalidSigner.selector, user3, user1)
-        );
+        vm.expectRevert(abi.encodeWithSelector(ERC2612InvalidSigner.selector, user3, user1));
         judgeToken.permit(user1, user2, value, deadline, v, r, s);
     }
 
@@ -589,20 +461,13 @@ contract JudgeTokenTest is Test {
 
         bytes32 structHash = keccak256(
             abi.encode(
-                keccak256(
-                    "Delegation(address delegatee,uint256 nonce,uint256 expiry)"
-                ),
-                delegatee,
-                nonce,
-                expiry
+                keccak256("Delegation(address delegatee,uint256 nonce,uint256 expiry)"), delegatee, nonce, expiry
             )
         );
 
         bytes32 domainSeparator = judgeToken.DOMAIN_SEPARATOR();
 
-        bytes32 digest = keccak256(
-            abi.encodePacked("\x19\x01", domainSeparator, structHash)
-        );
+        bytes32 digest = keccak256(abi.encodePacked("\x19\x01", domainSeparator, structHash));
 
         (uint8 v, bytes32 r, bytes32 s) = vm.sign(pKey, digest);
